Fixed brokenness
Problem was that messages for _all_ possible keys were generated.  This,
essentially, encoded a non-deterministic brute force attack where all
possible keys were tried in parallel - obviously one would work, so
secret would be derived.  We only want to generate PTKs for keys the
attacker has learned - hence, we check for bitstrings on the channel.

The same _could_ be said for nonces, and make the line something like
c(nap) & c(nsta) & c(key), but I don't know if that would be a more apt
protocol description.  As discussed, the nonces are sent in plaintext,
so they aren't really necessary for security.